requirements.txt
streamlit
openai
python-dotenv
pypdf
python-docx
reportlab